Qucsator now always returns data solely comprised of ComplexNumbers.
Whether in AC or DC mode, qucsator.parse will parse the block
of text from QUCS as being made of complex numbers.

The DMM only cares about the real portion of this data, and so
can just get the real portion from the qucsator data. The
oscilloscope will use both parts of the data, of course.

Update all tests.