Prevent component from being placed with leads too close
This fixes a bug where a component could be placed with
leads so close beneath component that they were impossible
to pick up and move again. Or, in the extreme case, such
as when a component was moved completely off the board,
the two leads would be placed on a single hole directly
beneath the component.

This now fixes the bug by having a component-specific
minimum distance, and moving the leads apart until this
minimum distance is satisfied.

[#47812435]