Add component drawer; allow new resistor to be added
If a section has the property "showComponentDrawer", the
Add Component button will be visible. Clicking on it will 
show the component drawer, which currently only shows a 
resistor.

This resistor can be dragged onto the breadboard, and a
resistor object will be created in the correct location.

The next commit will allow us to edit this component in
place.