Show in-place edit tooltip for new resistors
When a resistor is added, a tooltip appears with the
ability to edit the resistance or remove the component.

When the resistance is changed, we use the new view API
to update the resistor colors without need to add/remove
new DOM elements.

This edit view will become more generic when we add the
other component types.